class BOLIDLeTID
added function to calculate Ingrid Repins, Solar Energy 2020 % Deffects in each state (A, B, C). This currently solves the derivates for each state, but it the inputs still needs to account for irradiance, and temperature in the field to scale the attempt frequencies, and then use the outputs from this equation to calculate power loss.