HighPerformance
added new module "HighPerformance" designed to take an .ini file as input, scrape weather data from the NSRDBX data set on the HPC, and submit a job to the SLURM queue. It is extremely unstable, but has no impact on main.py. So far it is untested and is added only for development purposes, with extremely limited functionality.