Update weather.py
I had been using the meta information from "Source" to add meta information about the wind speed height but didn't realize that the meta information is not the same for all locations and this was causing crashing. I changed it to use the initial function call variable "database" to determine the source and the wind speed height.