Update weather.py
Created functions to do the mapping of meta and weather data to PVLib terms. Also eliminated the redundant instances of doing this. Updated a few of these terms to create a more comprehensive list. Created a function to make an *.csv file from weather and meta data. Created a function to read a generic *.csv file.