Update standards.py
Fixed some of the output statements to be clearer. Changed the poa_min default to b 400 W/m2 as opposed to 100 W/m2.

Lots of work to streamline the calculator for Xeff.  Removed superfluous passing of variables. Put in better default control over tilt to use latitude if none is provided. Also put in checks to make sure the tilt and azimuth are provided in the *.csv file.