Update weather.py
I added a function to get one meteorolgical data set by just specifying the longitude and latitude.

I also fixed some grammatical issues.

I also made it so that when it is calculating the RH it will display a notification and delete it when done. the notification was already there, I just made it delete it after.