Update weather.py
This adds in some functionality to provide a full set of meta data from another source to augment the PVGIS and NSRDB data. Doing this function call is not required and will default to not doing it for a geospatial analysis but default to do it for a single point look-up.