Update humidity.py
There were a lot of odd things wrong here but i restored it to a point of working with my code as it had before. I still need to go through and make sure the pytests still work.

There were also a number of functions that were duplicated so I deleted them. I think this happend in a bad merge.